handler: move package from /handler to /utils/handler
